// packages/ag-grid-community/src/interfaces/iRowPosition.ts
import type { RowPinnedType } from './iRowNode';

export interface RowPosition {
    /** A positive number from 0 to n, where n is the last row the grid is rendering
     * or -1 if you want to navigate to the grid header */
    rowIndex: number;

    /** Either 'top', 'bottom' or null/undefined (for not pinned) */
    rowPinned: RowPinnedType;
}